Use `has` for "Accept" header based routing
This allows the request `Accept` header to be parsed at the edge level
and route to an internal URL that informs the page of the requested
output format. Caching at the edge works as expected as well, allowing
us to use stale-while-revalidate in combination with `Accept` header
based routing.
